tcp server init

+++ rocket/net/io_thread_group.cc
@@ -4,7 +4,7 @@
 namespace rocket {
 IOThreadGroup::IOThreadGroup(int size) : m_size(size){
     m_io_thread_groups.resize(size);
-    for(size_t i = 0; i < m_size; i ++) {
+    for(int i = 0; i < m_size; i ++) {
         m_io_thread_groups[i] = new IOThread();
     }
 }
@@ -25,7 +25,7 @@ void IOThreadGroup::join() {
 }
 
 IOThread* IOThreadGroup::getIOThread() {
-    if(m_index == m_io_thread_groups.size() || m_index == -1) {
+    if(m_index == (int)m_io_thread_groups.size() || m_index == -1) {
         m_index = 0;
     }
     return m_io_thread_groups[m_index ++];
